+++ doc/changelog.rst
@@ -22,6 +22,13 @@ Version 0.6
   new tests, reaching 100% coverage, and several new gallery plots.
   (`PR 297 <https://github.com/fatiando/fatiando/pull/297>`__)
 
+**Bug fixes**
+
+* Fixes bug in ``fatiando.gravmag.imaging``, replacing
+  ``transform._getfreqs(x, y, data, shape)`` with
+  ``transform._fftfreqs(x, y, shape, shape)`` because that function was renamed.
+  (`PR 347 <https://github.com/fatiando/fatiando/pull/347>`__)
+
 
 .. _changelog-0.5: